Implement static type checking
* Introduct WheelMetadata type
* Move logic for analysing wheel metadata to seperate package
* Move requirement handling to seperate file
* Handle file:// urls
