+++ default.nix
@@ -1,5 +1,5 @@
 { stdenv, fetchurl, zip, makeWrapper, nix, nix-prefetch-git, nix-prefetch-hg
-, pythonPackages
+, pythonPackages, mypy
 , src ? { outPath = ./.; name = "pypi2nix"; }
 }:
 
@@ -27,6 +27,7 @@ in stdenv.mkDerivation rec {
   buildInputs = [
     pythonPackages.python pythonPackages.flake8
     zip makeWrapper nix.out nix-prefetch-git nix-prefetch-hg
+    mypy
   ];
   doCheck = true;
   sourceRoot = ".";
@@ -65,6 +66,7 @@ in stdenv.mkDerivation rec {
   '';
 
   checkPhase = ''
+    mypy ${src}/src --python-version 3.6
     flake8 ${src}/src
   '';
