+++ src/pypi2nix/utils.py
@@ -117,3 +117,10 @@ def prefetch_git(url, rev=None):
         )
     repo_data = json.loads(completed_proc.stdout)
     return repo_data
+
+
+def starts_with_any(s, prefixes):
+    for prefix in prefixes:
+        if s.startswith(prefix):
+            return True
+    return False
